requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
Markdown==3.7
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
